requirements.txt
mkdocs-material>=9.5
clickhouse-connect>=0.7
Jinja2>=3.1
pyyaml>=6.0
